refactor(common): extract ClipWord component
